+++ src/Dfc.CourseDirectory.Web/Helpers/ApprenticeshipProvisionHelper.cs
@@ -12,6 +12,7 @@ using Dfc.CourseDirectory.Services.Models;
 using Dfc.CourseDirectory.Services.Models.Apprenticeships;
 using Dfc.CourseDirectory.Services.Models.Regions;
 using Dfc.CourseDirectory.Services.VenueService;
+using Dfc.CourseDirectory.Web.ApprenticeshipBulkUpload;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
